# src/main/java/com/company/SocialNetwork/shared/PublicTimelineResponseDTO.java
package com.company.SocialNetwork.shared;

import com.company.SocialNetwork.timeline.TimelinePostDTO;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.HashSet;
import java.util.Set;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class PublicTimelineResponseDTO<T> {

    private Set<TimelinePostDTO> content = new HashSet<>();
    private String totalItems = "*";
    private Double nextScore;
}